% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amce.R
\name{amce}
\alias{amce}
\title{Estimator for the average causal mediation effect (AMCE)}
\usage{
amce(data, post, ..., cluster = NULL, subgroup = NULL, diff = NULL,
  subset = NULL)
}
\arguments{
\item{data}{Data}

\item{post}{Post treatment variable.}

\item{...}{The treatment components (without quotes).}

\item{cluster}{Variable indicating clusters if SEs should be
clustered.}

\item{subgroup}{Variable(s) indicating which subgroups, if any, to
independently calculate AMCEs for. If multiple subgroups are
provided, then it will iterate over each combination.}

\item{diff}{If \code{diff} is provided, it will calculate the
difference in AMCE between the unique values of \code{diff}.}

\item{subset}{Subset to estimate AMCEs for.}
}
\description{
Estimates the average causal mediation effect (AMCE) for all
provided treatments. Can estimate clustered and/or robust standard
errors.
}
\examples{
\dontrun{
 amce <- amce(data, post, treatment_1, treatment_2)
}
}
\keyword{AMCE}
