% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggsave2}
\alias{ggsave2}
\title{Plot both pdfs and pngs in one go.}
\usage{
ggsave2(plot, filename, ...)
}
\arguments{
\item{plot}{A ggplot2 plot object.}

\item{filename}{filename (without extenstion!)}

\item{...}{Other arguments passed to \code{ggsave()}.}
}
\description{
Plot both pdfs and pngs in one go.
}
