% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_diff_by}
\alias{plot_diff_by}
\title{Plot difference between closest and prefered candidates for
difference subgroups.}
\usage{
plot_diff_by(res, by)
}
\arguments{
\item{res}{results from \code{amce()} passed through
\code{add_labels()}.}

\item{by}{Variable to facet by.}
}
\description{
Plot difference between closest and prefered candidates for
difference subgroups.
}
