% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_diff_cntry_by}
\alias{plot_diff_cntry_by}
\title{Plot difference between closest and prefered candidates for
difference subgroups for each country.}
\usage{
plot_diff_cntry_by(res, by, by_lab)
}
\arguments{
\item{res}{results from \code{amce()} passed through
\code{add_labels()}.}

\item{by}{Variable to facet by.}

\item{by_lab}{Label for the \code{by} in the shape legend.}
}
\description{
Plot difference between closest and prefered candidates for
difference subgroups for each country.
}
