% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sub_by}
\alias{plot_sub_by}
\title{Plot AMCEs for closest and prefered candidates separately for
difference subgroups.}
\usage{
plot_sub_by(res, by)
}
\arguments{
\item{res}{results from \code{amce()} passed through
\code{add_labels()}.}

\item{by}{Variable to facet by.}
}
\description{
Plot AMCEs for closest and prefered candidates separately for
difference subgroups.
}
